requirements.txt
Flask==2.3.2
flask-bootstrap==3.3.7.1
dnspython==2.4.2
dkimpy==1.1.5
pyspf==2.0.14
requests==2.31.0
whois==0.9.10
langdetect==1.0.9
nltk==3.8.1
pandas==2.2.1
matplotlib==3.8.3
